Fixed errors due to merging OCR interfaces. You can pull now!

+++ BillSplit/src/billsplit/engine/Item.java
@@ -6,7 +6,7 @@ public class Item {
 	private String category;
 	private String details;
 	
-	public Item(String name, float cost) {
+	public Item(String name, double cost) {
 		this.name = name;
 		this.cost = cost;
 	}
